adicionado tabela de lanche

+++ src/main/resources/db.migration/V4__create-table-lanche.sql
@@ -1,5 +1,5 @@
 create table lanche(
-    id long SERIAL primary key,
+    id int SERIAL primary key,
     nome varchar(150) not null,
     descricao TEXT not null,
     preco decimal not null,
